Initial Programming
This is the completed product. The code was not testable much before
this update, so not much information is lost by only updating now. The
main things are a swap in X and Y coordinates relative to the previous
assignment, and fixing a few infinite recursive loops that were a result
of not having the code complete the first time it was run.