Update File Name and Added Complete Comments
Changed the file name to reflect which program number this is in the
class. Also added comments on literally everything, in case someone
would like to learn what my code does, or in case I need to learn what
my code does in the future. Finally, made the program backwards
compatible with the previous assignment. That brings me greater joy than
it probably should, but it's a cool thing.